REFACT _ structure for Currency / FIX _  problem with load CurrencyNBU
