+++ src/components/HeaderNav/CurrencyNav/helperCurrency.js
